[BACKPORT 2025.1] [#27324] YSQL: Preserve column order during backup restore of inherited tables
Summary:
Original commit: 6deb862d0affa6124c446098d94ec944405ffaae / D44232

In a scenario like

```
create table par(pc1 int);
create table ch(chc1 int) inherits (par);
alter table par add column pc2 int;
```
The order of columns in ch in both docdb and PG is `pc1, chc1, pc2`.

Such a table is currently output by `ysql_dump` during a backup as
```
create table par(pc1 int, pc2 int);
create table ch(chc1 int) inherits (par);
```
leading to the order of cols in ch being `pc1, pc2, chc1`.

This leads to a mismatch at DocDB snapshot import time when the docdb order does not match the PG order. DocDB snapshot import currently accepts "gaps" in column order to account for dropped columns but it does not accept columns in a different order.

While we could try to fix the docdb snapshot import to accept columns in a different order, it seems safer to imitate what PG binary upgrade does to preserve column order while dumping out the inheritance relationship. PG binary upgrade outputs the above as

```
create table par(pc1 int, pc2 int);
create table ch(pc1 int, chc1 int, pc2 int);
alter table ch inherits (par);

-- update pg_attribute attislocal = false for pc1, pc2 to reflect they are actually inherited
-- update pg_constraint.conislocal = false for inherited constraints
```

This change makes `ysql_dump --include-yb-metadata` behave similar to PG binary upgrade w.r.t the above. Some special cases:
1. PG also does this for dropped columns but we continue to ignore dropped columns in ysql_dump --include-yb-metadata as before, even for inherited children. We rely on the same docdb import snapshot code fixup for these.
2. For (stored) generated columns, PG outputs the generation expression separately for child tables in binary upgrade mode. Without this, it is not possible to `alter table .. inherit` the child as [PG15 does not allow generated columns to differ from parent to child](https://www.postgresql.org/docs/15/ddl-generated-columns.html). The same behavior is imitated for `include-yb-metadata`.
3. PG binary upgrade also outputs the fixup of pg_attribute `attislocal = false` for partition tables. However, we skip this because attislocal is already false after an `attach partition` of the child partition and we don't want to mess with existing partition flows.
Jira: DB-16834

Test Plan: Added a new unit test to the backup restore test that includes default values and generated columns. Existing tests in dump that cover partitions do not show a change in output.

Reviewers: fizaa, myang, loginov, mihnea

Reviewed By: mihnea

Subscribers: mihnea, yql

Differential Revision: https://phorge.dev.yugabyte.com/D44548

+++ src/postgres/src/bin/pg_dump/pg_dump.c
@@ -9391,6 +9391,10 @@ getTableAttrs(Archive *fout, TableInfo *tblinfo, int numTables)
  *
  * This function exists because there are scattered nonobvious places that
  * must be kept in sync with this decision.
+ *
+ * YB :  follow the same process for backup dumps of inheritance child table columns
+ * except for dropped columns, which we ignore because docdb snapshot import can
+ * tolerate dropped column differences.
  */
 bool
 shouldPrintColumn(const DumpOptions *dopt, const TableInfo *tbinfo, int colno)
@@ -9399,7 +9403,9 @@ shouldPrintColumn(const DumpOptions *dopt, const TableInfo *tbinfo, int colno)
 		return true;
 	if (tbinfo->attisdropped[colno])
 		return false;
-	return (tbinfo->attislocal[colno] || tbinfo->ispartition);
+	return tbinfo->attislocal[colno] ||
+		tbinfo->ispartition ||
+		(tbinfo->numParents > 0 && dopt->include_yb_metadata);
 }
 
 
@@ -16677,10 +16683,14 @@ dumpTableSchema(Archive *fout, const TableInfo *tbinfo)
 					 * Not Null constraint --- suppress if inherited, except
 					 * if partition, or in binary-upgrade case where that
 					 * won't work.
+					 * YB: For backups, follow binary-upgrade mode
+					 * for inherited child tables to preserve col order.
 					 */
 					print_notnull = (tbinfo->notnull[j] &&
 									 (!tbinfo->inhNotNull[j] ||
-									  tbinfo->ispartition || dopt->binary_upgrade));
+									  tbinfo->ispartition ||
+									  dopt->binary_upgrade ||
+									  dopt->include_yb_metadata));
 
 					/*
 					 * Skip column if fully defined by reloftype, except in
@@ -16857,9 +16867,11 @@ dumpTableSchema(Archive *fout, const TableInfo *tbinfo)
 			/*
 			 * Emit the INHERITS clause (not for partitions), except in
 			 * binary-upgrade mode.
+			 * YB: For backups, follow binary-upgrade mode
+			 * in inherited child tables to preserve col order.
 			 */
 			if (numParents > 0 && !tbinfo->ispartition &&
-				!dopt->binary_upgrade)
+				!dopt->binary_upgrade && !dopt->include_yb_metadata)
 			{
 				appendPQExpBufferStr(q, "\nINHERITS (");
 				for (k = 0; k < numParents; k++)
@@ -16996,8 +17008,12 @@ dumpTableSchema(Archive *fout, const TableInfo *tbinfo)
 		 * and matviews, even though they have storage, because we don't
 		 * support altering or dropping columns in them, nor can they be part
 		 * of inheritance trees.
+		 *
+		 * YB: For backups, follow the same process as binary-upgrade
+		 * specifically for inherited child tables
+		 * in order to preserve column order, except for dropped columns.
 		 */
-		if (dopt->binary_upgrade &&
+		if ((dopt->binary_upgrade || dopt->include_yb_metadata) &&
 			(tbinfo->relkind == RELKIND_RELATION ||
 			 tbinfo->relkind == RELKIND_FOREIGN_TABLE ||
 			 tbinfo->relkind == RELKIND_PARTITIONED_TABLE))
@@ -17006,29 +17022,36 @@ dumpTableSchema(Archive *fout, const TableInfo *tbinfo)
 			{
 				if (tbinfo->attisdropped[j])
 				{
-					appendPQExpBufferStr(q, "\n-- For binary upgrade, recreate dropped column.\n");
-					appendPQExpBuffer(q, "UPDATE pg_catalog.pg_attribute\n"
-									  "SET attlen = %d, "
-									  "attalign = '%c', attbyval = false\n"
-									  "WHERE attname = ",
-									  tbinfo->attlen[j],
-									  tbinfo->attalign[j]);
-					appendStringLiteralAH(q, tbinfo->attnames[j], fout);
-					appendPQExpBufferStr(q, "\n  AND attrelid = ");
-					appendStringLiteralAH(q, qualrelname, fout);
-					appendPQExpBufferStr(q, "::pg_catalog.regclass;\n");
+					/*
+					 * For YB backups, we don't need to recreate dropped cols because
+					 * docdb snapshot import can handle such gaps in the col order.
+					 */
+					if (!dopt->include_yb_metadata)
+					{
+						appendPQExpBufferStr(q, "\n-- For binary upgrade, recreate dropped column.\n");
+						appendPQExpBuffer(q, "UPDATE pg_catalog.pg_attribute\n"
+										"SET attlen = %d, "
+										"attalign = '%c', attbyval = false\n"
+										"WHERE attname = ",
+										tbinfo->attlen[j],
+										tbinfo->attalign[j]);
+						appendStringLiteralAH(q, tbinfo->attnames[j], fout);
+						appendPQExpBufferStr(q, "\n  AND attrelid = ");
+						appendStringLiteralAH(q, qualrelname, fout);
+						appendPQExpBufferStr(q, "::pg_catalog.regclass;\n");
 
-					if (tbinfo->relkind == RELKIND_RELATION ||
-						tbinfo->relkind == RELKIND_PARTITIONED_TABLE)
-						appendPQExpBuffer(q, "ALTER TABLE ONLY %s ",
-										  qualrelname);
-					else
-						appendPQExpBuffer(q, "ALTER FOREIGN TABLE ONLY %s ",
-										  qualrelname);
-					appendPQExpBuffer(q, "DROP COLUMN %s;\n",
+						if (tbinfo->relkind == RELKIND_RELATION ||
+							tbinfo->relkind == RELKIND_PARTITIONED_TABLE)
+							appendPQExpBuffer(q, "ALTER TABLE ONLY %s ",
+											qualrelname);
+						else
+							appendPQExpBuffer(q, "ALTER FOREIGN TABLE ONLY %s ",
+											qualrelname);
+						appendPQExpBuffer(q, "DROP COLUMN %s;\n",
 									  fmtId(tbinfo->attnames[j]));
+					}
 				}
-				else if (!tbinfo->attislocal[j])
+				else if (!tbinfo->attislocal[j] && (IsYugabyteEnabled && !tbinfo->ispartition))
 				{
 					appendPQExpBufferStr(q, "\n-- For binary upgrade, recreate inherited column.\n");
 					appendPQExpBufferStr(q, "UPDATE pg_catalog.pg_attribute\n"
@@ -17081,7 +17104,7 @@ dumpTableSchema(Archive *fout, const TableInfo *tbinfo)
 				}
 			}
 
-			if (OidIsValid(tbinfo->reloftype))
+			if (OidIsValid(tbinfo->reloftype) && !dopt->include_yb_metadata)
 			{
 				appendPQExpBufferStr(q, "\n-- For binary upgrade, set up typed tables this way.\n");
 				appendPQExpBuffer(q, "ALTER TABLE ONLY %s OF %s;\n",
